UI (#123)
* merging

* sign up progress

* sign up ui

* call and edit profile

* email and ride details fragment

* request fragment

* removed unnecessary classesA

* waiting for driver fragment

* profile, call, email

+++ app/src/main/java/ca/ualberta/boost/models/Driver.java
@@ -42,7 +42,7 @@ public class Driver extends User {
      * @return
      *      Returns the number of thumbs up ratings the driver has
      */
-    public double getPositiveRating() {
+    public int getPositiveRating() {
         return thumbsUp;
     }
